style: fix file arrangements
