+++ src/engine/index.ts
@@ -1,5 +1,6 @@
 export { compile } from "./compiler/compiler";
 export { converter } from "./compiler/converter";
 export { render, renderFile } from "./utils/render";
-export { FrostError } from "./utils/FrostError";
+export { FrostError } from "./utils/error";
+
 export * as middleware from "./utils/middleware";
