--- src/minifier/js.ts
@@ -1,3 +1,0 @@
-import { minify } from "terser";
-
-export default async (text: string) => (await minify(text)).code;
